slack-bot/requirements.txt
slack-bolt==1.18.1
slack-sdk==3.26.1
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
